+++ app/templates/config/webpack.es6.js
@@ -1,0 +1,28 @@
+const merge = require('webpack-merge');
+const common = require('./webpack.common');
+
+module.exports = merge(common, {
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            presets: [
+              [
+                '@babel/preset-env',
+                {
+                  targets: {
+                    esmodules: true,
+                  },
+                  debug: false,
+                },
+              ],
+            ],
+          },
+        },
+      },
+    ],
+  },
+});