+++ app/templates/js/_main.js
@@ -1,5 +1,0 @@
-const foo = 'bar';
-
-const baz = () => foo;
-
-baz();